requirements.txt
beautifulsoup4==4.10.0
bidict==0.22.0
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
netifaces==0.10.6
pyretrommo==0.0.1
python-engineio==4.3.4
python-socketio==5.7.1
requests==2.26.0
soupsieve==2.3.2.post1
urllib3==1.26.11
websocket-client==1.3.3
